requirements.txt
requests>=2.32.0,<3
psycopg2-binary>=2.9,<3
python-dotenv>=1.1.0,<2
python-telegram-bot>=21,<22
fastapi>=0.115,<1
uvicorn[standard]>=0.35,<1
email-validator>=2.1,<3
cachetools>=5.3,<6
langgraph>=1.0.1,<2
langchain-google-genai>=2.1.12,<3
langchain-google-vertexai>=2.1.0,<3
